Update MaterialManager.java
Materials are stored as MaterialData in materials.
Aliases are stored as strings in aliases.
Materials are loaded in loadMaterials()
addMaterial is a temporary test function for spawning stuff in. Need to figure out how to spawn potions and stuff.
getMaterialPrice for getting prices.
removed deep scan from loadAliases which could cause issues if config was malformed.
removed deep scan from loadFile.