Update Math.java
Fixed #49
I've made some slight adjustment to the calculatePrice to fix getPrice.
getPrice is default broken.

I've also added some logging if an error does occur, in regards to extreme double values causing negative results, which prints the error type, which should hopefully help with debugging later if this problem still exists, which it may not.