1.18.3 RC1
Economy rework in progress, release in 1.18.3 RC2. Which will also be jar'd and shipped.

Fixed "Preferred Provider" setting not being obeyed properly. Issue 55 can now be fixed through the use of a different economy provider. Such as setting "Preferred Provider" to "Essentials" or another provider. Providers can be found on the startup screen for DivinityEconomy.

Fixed config not being updated when values are missing.

Added new setting "Ignore Save Messages" which allows the user to disable the x.yml was saved message that occurs every 10 minutes, by default.

Added new settings for setting the maximum and minimum values of items. Disable with any negative value, for example -1. By default, there is no cap on the maximum value of an item. By default, the cap for the minimum value for an item is 0.001.

41 changed files due to name changes.